+++ changelog.md
@@ -4,7 +4,7 @@
 * Added Better Beacon
   * A new beacon with more effects, more amplifiers but with a time cost
 * Added unbreakable items
-  * Items will no longer break, instead they stay at one durability. Doesn't work on shears, fishing rods, brush, flint and steel, bow, crossbow, shield, trident,  
+  * Items will no longer break, instead they stay at one durability.  
 * Added Name Tags feature
   * Retrieve name tags from name tagged entities, from changing their name to their demise
 * Seasons